Split code into different files
Remove prepareSendgridMailV3 func
