+++ mailing/sendgrid.go
@@ -3,8 +3,6 @@ package mailing
 import (
 	"context"
 	"fmt"
-	"github.com/gazebo-web/gz-go/v8"
-	"github.com/gazebo-web/gz-go/v8/structs"
 	"github.com/sendgrid/sendgrid-go/helpers/mail"
 	"net/http"
 )
@@ -18,10 +16,7 @@ type sendgridEmailService struct {
 }
 
 func (s *sendgridEmailService) emailBuilder() sendgridEmailBuilder {
-	return sendgridEmailBuilder{
-		personalization: mail.NewPersonalization(),
-		mail:            mail.NewV3Mail(),
-	}
+	return newSendgridEmailBuilder()
 }
 
 // Send sends an email from sender to the given recipients. The email body is composed by an HTML template
@@ -55,65 +50,23 @@ func (s *sendgridEmailService) Send(ctx context.Context, sender string, recipien
 	return nil
 }
 
-// sendgridEmailBuilder allows building mail.SGMailV3 emails.
-type sendgridEmailBuilder struct {
-	personalization *mail.Personalization
-	mail            *mail.SGMailV3
+// NewSendgridEmailSender initializes a new Sender with a sendgrid client. It will send emails using Go templates.
+// See NewTemplateSender for conveniently handling Go templates in your project.
+func NewSendgridEmailSender(client sendgridSender) Sender {
+	return newSendgridEmailSenderWithContentInjector(client, injectHTMLContent)
 }
 
-// Sender sets the email address where the resulting email comes from.
-func (b sendgridEmailBuilder) Sender(from string) sendgridEmailBuilder {
-	b.personalization.AddFrom(mail.NewEmail("", from))
-	return b
+// NewSendgridDynamicTemplatesEmailSender initializes a new Sender with a sendgrid client. It will send emails through
+// sendgrid using dynamic templates defined in the Sendgrid dashboard.
+func NewSendgridDynamicTemplatesEmailSender(client sendgridSender) Sender {
+	return newSendgridEmailSenderWithContentInjector(client, injectTemplateContent)
 }
 
-// Recipients sets the email addresses where the resulting email will be sent to.
-func (b sendgridEmailBuilder) Recipients(to []string) sendgridEmailBuilder {
-	b.personalization.AddTos(parseSendgridEmails(to)...)
-	return b
-}
-
-// CC sets the email addresses where the resulting email will be carbon-copied to.
-func (b sendgridEmailBuilder) CC(ccs []string) sendgridEmailBuilder {
-	b.personalization.AddCCs(parseSendgridEmails(ccs)...)
-	return b
-}
-
-// BCC sets the email addresses where the resulting email will blind carbon-copied to.
-func (b sendgridEmailBuilder) BCC(bccs []string) sendgridEmailBuilder {
-	b.personalization.AddBCCs(parseSendgridEmails(bccs)...)
-	return b
-}
-
-// Subject sets the resulting email's subject.
-func (b sendgridEmailBuilder) Subject(subject string) sendgridEmailBuilder {
-	b.personalization.Subject = subject
-	return b
-}
-
-// Content sets the resulting email's body with the respective content type.
-// It's mutually exclusive with Template.
-func (b sendgridEmailBuilder) Content(contentType string, content string) sendgridEmailBuilder {
-	b.mail.AddContent(mail.NewContent(contentType, content))
-	return b
-}
-
-// Template defines the dynamic template identified by ID, and passes the given data as the template
-// parameters.
-func (b sendgridEmailBuilder) Template(id string, data any) sendgridEmailBuilder {
-	var err error
-	b.personalization.DynamicTemplateData, err = structs.ToMap(data)
-	if err != nil {
-		return b
+func newSendgridEmailSenderWithContentInjector(client sendgridSender, injector contentInjector) Sender {
+	return &sendgridEmailService{
+		client:          client,
+		contentInjector: injector,
 	}
-	b.mail.SetTemplateID(id)
-	return b
-}
-
-// Build creates the email from all the parameters previously used.
-func (b sendgridEmailBuilder) Build() *mail.SGMailV3 {
-	b.mail.AddPersonalizations(b.personalization)
-	return b.mail
 }
 
 // parseSendgridEmails converts the given slice of emails to sendgrid emails.
@@ -123,39 +76,4 @@ func parseSendgridEmails(emails []string) []*mail.Email {
 		out[i] = mail.NewEmail("", emails[i])
 	}
 	return out
-}
-
-// contentInjector defines the function signature that injects content injection into sendgridEmailBuilder
-type contentInjector func(b sendgridEmailBuilder, key string, data any) (sendgridEmailBuilder, error)
-
-// injectTemplateContent injects the information needed to send a Sendgrid email using dynamic templates.
-func injectTemplateContent(b sendgridEmailBuilder, id string, data any) (sendgridEmailBuilder, error) {
-	return b.Template(id, data), nil
-}
-
-// injectHTMLContent injects the actual content after parsing an HTML Go template.
-func injectHTMLContent(b sendgridEmailBuilder, path string, data any) (sendgridEmailBuilder, error) {
-	content, err := gz.ParseHTMLTemplate(path, data)
-	if err != nil {
-		return b, err
-	}
-	return b.Content("text/html", content), nil
-}
-
-// NewSendgridEmailSender initializes a new Sender with a sendgrid client. It will send emails using Go templates.
-// See NewTemplateSender for conveniently handling Go templates in your project.
-func NewSendgridEmailSender(client sendgridSender) Sender {
-	return &sendgridEmailService{
-		client:          client,
-		contentInjector: injectHTMLContent,
-	}
-}
-
-// NewSendgridDynamicTemplatesEmailSender initializes a new Sender with a sendgrid client. It will send emails through
-// sendgrid using dynamic templates defined in the Sendgrid dashboard.
-func NewSendgridDynamicTemplatesEmailSender(client sendgridSender) Sender {
-	return &sendgridEmailService{
-		client:          client,
-		contentInjector: injectTemplateContent,
-	}
 }
